requirements.txt
jsonschema==3.1.1
pyaml==19.4.1
pyserial